requirements.txt
streamlit
pyngrok
keras
tensorflow-cpu
datetime
joblib
numpy
pandas
scikit-learn
plotly
